+++ src/main/groovy/nextflow/validation/validators/evaluators/SchemaEvaluator.groovy
@@ -1,5 +1,6 @@
 package nextflow.validation.validators.evaluators
 
+import org.json.JSONObject
 import dev.harrel.jsonschema.Evaluator
 import dev.harrel.jsonschema.EvaluationContext
 import dev.harrel.jsonschema.JsonNode
@@ -12,6 +13,7 @@ import static nextflow.validation.utils.Common.getBasePath
 import static nextflow.validation.utils.Files.fileToJson
 import nextflow.validation.config.ValidationConfig
 import nextflow.validation.validators.JsonSchemaValidator
+import nextflow.validation.validators.ValidationResult
 
 /**
  * @author : nvnieuwk <nicolas.vannieuwkerke@ugent.be>
@@ -54,8 +56,8 @@ class SchemaEvaluator implements Evaluator {
         def Object json = fileToJson(file, Path.of(schemaFull))
         def validator = new JsonSchemaValidator(config)
 
-        def Tuple2<List<String>,List<String>> validationResult = validator.validate(json, schemaFull)
-        def validationErrors = validationResult[0]
+        def ValidationResult validationResult = validator.validate(json, schemaFull)
+        def List<String> validationErrors = validationResult.getErrors((json instanceof JSONObject) ? "parameter" : "field")
         if (validationErrors) {
             def List<String> errors = ["Validation of file failed:"] + validationErrors.collect { "\t${it}" as String}
             return Evaluator.Result.failure(errors.join("\n"))